Major overhaul of WebSockets.js
This is a complete rewrite of WebSockets.js. I started completely over. The old code used awful globals, had really awful syntax and bad commenting, and was just generally a mess that wasn't worth salvaging. It worked, but if we ever had to modify it, it would've been hell. This rewrite turns that into a pleasant, object-ish new set of code that's got much better comments.

As of this commit, ZephyrCab as a whole is not functional; it's in the middle of a massive overhaul on the ui.js file, which is deeply linked to sim.js and consequently none of the actual simulation works right due to variables being shifted around and undefined. This commit is by no means a stable application. However, websockets.js, which is the only focus of this commit, is entirely stable as far as I can tell. There may be some work on the decoder objects required, but this file itself should be totally solid.