Creating temporary sim-new.js for major overhaul
I plan on rewriting a large portion of sim.js, mostly for readability and conciseness. The current code is functional, but it's very very ugly and difficult to understand and debug. I created sim-new.js as a temporary place to write the new sim.accel(), and that way I can work off the old sim.accel(). Eventually, of course, the contents of sim-new.js will be integrated into the real sim.js.