Adding JMRI Jetty autoconnection
In this commit I added the ability to automatically connect to JMRI without settings.json editing. By default, ZephyrCab assumes it's running on JMRI's built-in Jetty web server. When you access ZephyrCab, it will examine the location.hostname and location.port values and try to make a WebSockets connection to the same place.

There is an option to disable this, documented in comments in settings.json.