Fixed resending issue in decoders.js, disabling fuel codebase, more work on sim.js/ui.js overhaul
I fixed a serious issue with my LokSound Select decoder file in decoders.js. When you call whatever.set(boolean), it sends a WebSockets command to JMRI EVERY TIME! This doesn't sound like a flaw, but when you've got a physics engine cycling every 100ms, these commands can actually bring JMRI to a crawl. I used the existing whatever.state variables in the decoder object spec and put an IF statement in the function .set() methods to prevent unnecessary network traffic.

I also worked a little more on sim.js and ui.js, and I commented out all the executable code relating to fuel. I left all the variables and whatnot inside bundles.json, but since fuel usage/engine shutoff is a very high maintenance and low priority feature, I've decided to get rid of it entirely during the further development of the rest of the physics engine. Later on once I've got a basic physics engine down, I'll come in and rebuild it, but it's too much of a headache for the time being.