Repairing massive bug with for loop variable scopes
The main for loop in sim.js had a serious issue with a for loop nested inside of it; their i variables were overlapping somehow. I'm not really sure how, since I declared both in a local scope (I think) but changing the nested loop's variable to x instead of i fixed it, so that's what I did.