Adding bail-off button and color change system
I've changed the bail off checkbox to a button, which I believe is more realistic. However I did add a color change function (at ui.bailoff.set(boolean) ) that I plan to use to turn the button red when it's possible to bail off. So when the user applies the automatic brake, they'll be reminded that they need to push the "bail off" button.