Removing frivolous air stuff
I removed the "needsEngine" option for the compressor because it's
simply a pain and it isn't even implemented right now, so no sense
cluttering up the code. I also removed all the information about the
flow rate for the bell, because it's so relatively small it's not worth
wasting CPU on.