Work on sim, BROKEN
This is just a messy commit so I can access from other PCs. The locomotive portion of the sim works fine but the rolling stock portion is very messed up. The rolling resistance doesn't become 0 when the train stops, and it begins to pull the train backwards. MAJOR zero crossing issues in this commit.