Cleaning up logging system
I made a nice wrapper for console.log. It's inside scripts/pretty-logs.js, and the Object log. There are a handful of categories with predefined keywords to make searching/filtering the log easier. Upon startup, ZephyrCab dumps all the keyword info to the console in case you forget.

This resolves #34