First iteration of setup data file generator
This version of the setup page just generates some nicely formatted JSON, which users will have to copy and paste into their bundles file. It's also not able to include the functions necessary for the simulator to actually work, so some additional code needs to be written so that zephyrcab is smart enough to add those functions automagically if they are undefined. However I am committing this first as the aforementioned code may become unnecessary if I choose to leverage the JMRI web server's directory listing feature to allow users to store bundles in individual files in a folder.