Creating epic new bundle generator
This should be a fully functional generator for locomotive bundles. This also includes some MASSIVE changes to the way bundles are stored. They are now stored in individual .zephyrcab files, inside the /cfg/bundles directory. You must put the name of the file in an array inside /cfg/bundles.list.json, but once you do that ZephyrCab will load each bundle automagically. It will add them to bundles.locomotives AND it will make sure the bundles have the necessary functions (as the new generator integrated into ZephyrCab generates only the static data, not the functions).